Refactor text area and feedback
