+++ __tests__/html2/feedbackForm/feedback.form.scroll.fluent.html
@@ -1,0 +1,10 @@
+<!doctype html>
+<html lang="en-US">
+  <head>
+    <title>Feedback Form (fluent): Scrollbar behavior when input is long</title>
+    <script>
+      location.href = './feedback.form.scroll?theme=fluent';
+    </script>
+  </head>
+  <body></body>
+</html>
