+++ packages/component/src/TextArea/index.ts
@@ -1,0 +1,1 @@
+export { default as TextArea } from './TextArea';
