--- packages/fluent-theme/src/components/sendBox/TextArea.module.css
@@ -1,82 +1,0 @@
-:global(.webchat-fluent) .sendbox__text-area {
-  display: grid;
-  grid-template-areas: 'main';
-  max-height: 200px;
-  overflow: auto;
-  scrollbar-gutter: stable;
-}
-
-:global(.webchat-fluent) .sendbox__text-area--hidden {
-  /* TODO: Not perfect way of hiding the text box. */
-  height: 0;
-  visibility: collapse;
-}
-
-:global(.webchat-fluent) .sendbox__text-area--in-completion {
-  .sendbox__text-area-doppelganger {
-    visibility: unset;
-  }
-
-  .sendbox__text-area-input {
-    background-color: transparent;
-    caret-color: var(--webchat-colorNeutralForeground1);
-    color: transparent;
-  }
-
-  textarea::placeholder {
-    color: transparent;
-  }
-}
-
-:global(.webchat-fluent) .sendbox__text-area-shared {
-  border: none;
-  font: inherit;
-  grid-area: main;
-  outline: inherit;
-  overflow-wrap: anywhere;
-  resize: inherit;
-}
-
-:global(.webchat-fluent) .sendbox__text-area-doppelganger {
-  overflow: visible;
-  pointer-events: none;
-  user-select: none;
-  visibility: hidden;
-  white-space: pre-wrap;
-}
-
-:global(.webchat-fluent) .sendbox__text-area-input {
-  background-color: inherit;
-  color: currentColor;
-  contain: size;
-  overflow: hidden;
-  padding: 0;
-}
-
-:global(.webchat-fluent) .sendbox__text-area--scroll {
-  /* Edge uses -webkit-scrollbar if scrollbar-* is not set */
-  scrollbar-color: unset;
-  scrollbar-width: unset;
-  /* Firefox */
-  -moz-scrollbar-color: var(--webchat-colorNeutralBackground5) var(--webchat-colorNeutralForeground2);
-  -moz-scrollbar-width: thin;
-
-  /* Chrome, Edge, and Safari */
-  &::-webkit-scrollbar {
-    width: 8px;
-  }
-
-  &::-webkit-scrollbar-track {
-    background-color: var(--webchat-colorNeutralBackground5);
-    border-radius: 16px;
-  }
-
-  &::-webkit-scrollbar-thumb {
-    background-color: var(--webchat-colorNeutralForeground2);
-    border-radius: 16px;
-  }
-
-  &::-webkit-scrollbar-corner {
-    background-color: var(--webchat-colorNeutralBackground5);
-  }
-}
