+++ packages/fluent-theme/src/components/theme/Theme.module.css
@@ -225,11 +225,11 @@
 
 /* Scrollbars */
 :global(.webchat-fluent).theme :global(.send-box-attachment-bar.send-box-attachment-bar--as-list-item),
+:global(.webchat-fluent).theme :global(.text-area.text-area--scroll),
 :global(.webchat-fluent).theme :global(.webchat__basic-transcript .webchat__basic-transcript__scrollable),
-:global(.webchat-fluent).theme :global(.webchat__view-code-dialog__code-body),
 :global(.webchat-fluent).theme :global(.webchat__render-markdown [data-math-type='block']),
-:global(.webchat-fluent).theme :global(.webchat__feedback-form-text-area-input--scroll),
-:global(.webchat-fluent).theme :global(.webchat__view-code-dialog__body) {
+:global(.webchat-fluent).theme :global(.webchat__view-code-dialog__body),
+:global(.webchat-fluent).theme :global(.webchat__view-code-dialog__code-body) {
   /* Edge uses -webkit-scrollbar if scrollbar-* is not set */
   scrollbar-color: unset;
   scrollbar-width: unset;
@@ -252,7 +252,12 @@
     width: var(--webchat-spacingVerticalMNudge);
   }
 
-  &:hover::-webkit-scrollbar-thumb {
+  &::-webkit-scrollbar-thumb {
+    background-color: transparent;
+  }
+
+  &:hover::-webkit-scrollbar-thumb,
+  &:has(> :global(.text-area-input:focus))::-webkit-scrollbar-thumb {
     background-clip: padding-box;
     background-color: var(--webchat-colorNeutralForeground5);
     border-radius: 10px;
@@ -625,35 +630,31 @@
 }
 
 /* Feedback Form */
-/* TODO: Need review */
-:global(.webchat-fluent).theme :global(.webchat__feedback-form) {
-  :global(.webchat__feedback-form__form-header) {
+:global(.webchat-fluent).theme :global(.feedback-form) {
+  :global(.feedback-form__form-header) {
     color: var(--webchat-colorNeutralForeground1);
-    font-family: var(--webchat-fontFamilyBase);
     font-size: var(--webchat-fontSizeBase300);
     font-weight: var(--webchat-fontWeightRegular);
     line-height: var(--webchat-lineHeightBase300);
   }
 
-  :global(.webchat__feedback-form__form) {
+  :global(.feedback-form__form) {
     gap: var(--webchat-spacingVerticalXS);
   }
 
-  :global(.webchat__feedback-form__form-footer) {
+  :global(.feedback-form__form-footer) {
     color: var(--webchat-colorNeutralForeground1);
-    font-family: var(--webchat-fontFamilyBase);
     font-size: var(--webchat-fontSizeBase100);
     font-weight: var(--webchat-fontWeightRegular);
     line-height: var(--webchat-lineHeightBase200);
   }
 
-  :global(.webchat__feedback-form__submit-button) {
+  :global(.feedback-form__submit-button) {
     background-color: var(--webchat-colorBrandBackground);
     border: var(--webchat-strokeWidthThin) solid var(--webchat-colorBrandBackground);
     border-radius: var(--webchat-borderRadiusMedium);
     color: var(--webchat-colorNeutralForegroundOnBrand);
     cursor: pointer;
-    font-family: var(--webchat-fontFamilyBase);
     font-size: var(--webchat-fontSizeBase300);
     padding: var(--webchat-spacingVerticalNone) var(--webchat-spacingHorizontalS);
     transition: background-color var(--webchat-durationNormal) var(--webchat-curveDecelerateMid);
@@ -674,13 +675,12 @@
     }
   }
 
-  :global(.webchat__feedback-form__cancel-button) {
+  :global(.feedback-form__cancel-button) {
     background-color: var(--webchat-colorNeutralBackground1);
     border: var(--webchat-strokeWidthThin) solid var(--webchat-colorNeutralStroke1);
     border-radius: var(--webchat-borderRadiusMedium);
     color: var(--webchat-colorNeutralForeground1);
     cursor: pointer;
-    font-family: var(--webchat-fontFamilyBase);
     font-size: var(--webchat-fontSizeBase300);
     padding: var(--webchat-spacingVerticalNone) var(--webchat-spacingHorizontalS);
     transition: background-color var(--webchat-durationNormal) var(--webchat-curveDecelerateMid);
@@ -688,6 +688,7 @@
     &:hover {
       background-color: var(--webchat-colorNeutralBackground1Hover);
       border: var(--webchat-strokeWidthThin) solid var(--webchat-colorNeutralStroke1Hover);
+      color: currentColor;
     }
 
     &:focus-visible {
@@ -701,21 +702,20 @@
     }
   }
 
-  :global(.webchat__feedback-form-text-area) {
+  :global(.feedback-form__text-box) {
     background-color: var(--webchat-colorNeutralBackground1);
     border-radius: var(--webchat-borderRadiusMedium);
     border: var(--webchat-strokeWidthThin) solid var(--webchat-colorNeutralStroke1);
-    font-family: var(--webchat-fontFamilyBase);
     font-size: var(--webchat-fontSizeBase300);
     line-height: var(--webchat-lineHeightBase300);
     gap: var(--webchat-spacingVerticalSNudge);
     padding: var(--webchat-spacingVerticalS) var(--webchat-spacingHorizontalM);
+
     &::after {
       border-bottom-left-radius: var(--webchat-borderRadiusMedium);
       border-bottom-right-radius: var(--webchat-borderRadiusMedium);
       border-bottom: var(--webchat-strokeWidthThicker) solid var(--webchat-colorBrandForeground1);
       clip-path: inset(calc(100% - var(--webchat-strokeWidthThicker)) 50% 0 50%);
-      transition: clip-path var(--webchat-durationNormal) var(--webchat-curveDecelerateMid);
     }
     &:focus-within {
       border: var(--webchat-strokeWidthThin) solid var(--webchat-colorNeutralStroke1);
